+++ src/pytorch_lightning/plugins/precision/sharded_native_amp.py
@@ -15,7 +15,7 @@ from typing import Optional, Union
 
 from typing_extensions import Literal
 
-from lightning_fabric.strategies.fairscale import _FAIRSCALE_AVAILABLE
+from pytorch_lightning.overrides.fairscale import _FAIRSCALE_AVAILABLE
 from pytorch_lightning.plugins.precision.native_amp import MixedPrecisionPlugin
 from pytorch_lightning.utilities.exceptions import MisconfigurationException