+++ tests/tests_pytorch/trainer/optimization/test_optimizers.py
@@ -557,25 +557,6 @@ def test_invalid_optimizer_dict_raises(tmpdir):
         trainer.fit(model)
 
 
-def test_warn_invalid_scheduler_key_in_manual_optimization(tmpdir):
-    """Test warning when invalid scheduler keys are provided in manual optimization."""
-
-    class TestModel(BoringModel):
-        def __init__(self):
-            super().__init__()
-            self.automatic_optimization = False
-
-        def configure_optimizers(self):
-            opt = optim.SGD(self.layer.parameters(), lr=0.1)
-            sch = optim.lr_scheduler.StepLR(opt, step_size=1)
-            return [opt], [{"scheduler": sch, "interval": "epoch"}]
-
-    model = TestModel()
-    trainer = Trainer(default_root_dir=tmpdir, fast_dev_run=True)
-    with pytest.warns(RuntimeWarning, match="the keys will be ignored"):
-        trainer.fit(model)
-
-
 @RunIf(min_cuda_gpus=2, standalone=True)
 def test_optimizer_state_on_device(tmpdir):
     """Test that optimizers that create state initially at instantiation still end up with the state on the GPU."""